Typed lambda events and moved slack-auth to shared
