+++ package.json
@@ -8,11 +8,13 @@
     "lint": "tslint --fix \"src/**/*.ts\" --project tsconfig.json",
     "test": "yarn test:unit",
     "test:unit": "jest src",
-    "test:integration": "jest integration-tests"
+    "test:integration": "jest integration-tests",
+    "tsc": "tsc"
   },
   "author": "Juho Friman <juho.friman@solita.fi>",
   "license": "MIT",
   "devDependencies": {
+    "@types/aws-lambda": "^8.10.39",
     "@types/crypto-js": "^3.1.43",
     "@types/jest": "^24.0.13",
     "@types/moment-timezone": "^0.5.12",
