+++ tsconfig.json
@@ -11,8 +11,8 @@
         "removeComments": false,
         "preserveConstEnums": true,
         "sourceMap": true,
-        "skipLibCheck": true,
 
+        "skipLibCheck": true,
         "typeRoots": [
             "./node_modules/@types"
         ],
